converted index.js to ts
